refactor(data): Add Debug and Clone derives to core data structures
Adds the #[derive(Debug)] and #[derive(Clone)] traits to all primary
state, configuration, and result structs (State, RobotConfig,
ControlCommand, etc.).

This significantly improves the usability and robustness of the module
by allowing:
- Easy printing and debugging of complex objects.
- Safe, deep copying of state and configuration data across function
  calls or threads.
